requirements.txt
openai==1.57.0
playwright==1.49.0
pytest==8.3.4
pytest-playwright==0.6.2
requests==2.32.3
PyGithub==2.5.0
PyYAML==6.0.2
